+++ apps/ledger-live-desktop/.eslintrc.js
@@ -58,9 +58,6 @@ module.exports = {
         "plugin:@typescript-eslint/recommended",
       ],
       parser: "@typescript-eslint/parser",
-      parserOptions: {
-        project: "./src/tsconfig.json",
-      },
       rules: {
         "space-before-function-paren": 0,
         "no-prototype-builtins": 0,
