+++ apps/ledger-live-desktop/src/renderer/actions/modals.ts
@@ -1,4 +1,3 @@
-// @flow
 import { createAction } from "redux-actions";
 
 export const openModal = createAction("MODAL_OPEN", (name, data) => ({ name, data }));
