+++ libs/ledger-live-common/package.json
@@ -87,7 +87,7 @@
     "@ledgerhq/hw-transport-mocker": "workspace:^",
     "@ledgerhq/hw-transport-node-speculos": "workspace:^",
     "@ledgerhq/json-bignumber": "^1.1.0",
-    "@ledgerhq/live-app-sdk": "^0.6.1",
+    "@ledgerhq/live-app-sdk": "file:../../../live-app-sdk",
     "@ledgerhq/logs": "workspace:^",
     "@polkadot/types": "8.5.1",
     "@polkadot/types-known": "8.5.1",
